+++ tests/Koalesce.Tests/GlobalUsings.cs
@@ -11,3 +11,5 @@ global using Microsoft.Extensions.Options;
 global using System.Net;
 global using System.Net.Sockets;
 global using Koalesce.Core.Options;
+global using Koalesce.Options;
+global using Koalesce.Constants;