+++ tests/Koalesce.Tests/OptionsUnitTests.cs
@@ -4,6 +4,81 @@ namespace Koalesce.Tests;
 public class OptionsUnitTests : KoalesceUnitTestBase
 {
 	[Fact]
+	public void Koalesce_WhenValidConfiguration_ShouldBindKoalesceOptions()
+	{
+		// Arrange
+		var appSettingsStub = new
+		{
+			Koalesce = new
+			{
+				Info = new { Title = "My Koalesced API" },
+				MergedEndpoint = "/v1/mergedapidefinition.yaml",
+				Sources = new List<ApiSource>
+				{
+					new ApiSource { Url = "https://localhost:5001/v1/apidefinition.json" },
+					new ApiSource { Url = "https://localhost:5002/v1/apidefinition.json" }
+				}
+			}
+		};
+
+		var configuration = ConfigurationHelper
+			.BuildConfigurationFromObject(appSettingsStub);
+
+		Services.AddKoalesce(configuration);
+
+		var provider = Services.BuildServiceProvider();
+
+		var expectedRoutes = new List<ApiSource>()
+		{
+			new ApiSource { Url = "https://localhost:5001/v1/apidefinition.json" },
+			new ApiSource { Url = "https://localhost:5002/v1/apidefinition.json" }
+		};
+
+		// Act
+		var options = provider.GetService<IOptions<KoalesceOptions>>()?.Value;
+
+		// Assert
+		Assert.NotNull(options);
+		Assert.Equal("My Koalesced API", options.Info.Title);
+		Assert.Equal("/v1/mergedapidefinition.yaml", options.MergedEndpoint);
+
+		Assert.Equal(expectedRoutes.Count, options.Sources.Count);
+		Assert.Equal(expectedRoutes[0].Url, options.Sources[0].Url);
+		Assert.Equal(expectedRoutes[1].Url, options.Sources[1].Url);
+	}
+
+	[Fact]
+	public void Koalesce_WhenNonRequiredConfigValuesAreMissing_ShouldUseDefaultValues()
+	{
+		// Arrange
+		var appSettingsStub = new
+		{
+			Koalesce = new
+			{
+				MergedEndpoint = "/v1/mergedapidefinition.json",
+				Sources = new List<ApiSource>
+			{
+				new ApiSource { Url = "https://localhost:5001/v1/apidefinition.json" }
+			}
+			}
+		};
+
+		var configuration = ConfigurationHelper
+			.BuildConfigurationFromObject(appSettingsStub);
+
+		Services.AddKoalesce(configuration);
+
+		var provider = Services.BuildServiceProvider();
+
+		// Act
+		var options = provider.GetService<IOptions<KoalesceOptions>>()?.Value;
+
+		// Assert
+		Assert.NotNull(options);
+		Assert.Equal(KoalesceConstants.DefaultTitle, options.Info.Title); // Default title should be set
+	}
+
+	[Fact]
 	public void Koalesce_WithInvalidGatewayUrl_ShouldThrowValidationException()
 	{
 		// Arrange
@@ -33,7 +108,7 @@ public class OptionsUnitTests : KoalesceUnitTestBase
 			var options = provider.GetRequiredService<IOptions<Options.KoalesceOptions>>().Value;
 		});
 
-		Assert.Contains(Constants.KoalesceConstants.ApiGatewayBaseUrlValidationError, exception.Message);
+		Assert.Contains(KoalesceConstants.ApiGatewayBaseUrlValidationError, exception.Message);
 	}
 
 	[Fact]